Update installation.md
Nginx supports, but discourages the use of conditional statements in configuration files. The Nginx HTTPS server configuration was separated from the HTTP server configuration in order to remove the conditional statement that forced HTTPS to be enabled if the block was uncommented. Since the rules for the HTTP and HTTPS servers are now separated, `return` will only run once, instead of entering into an infinite loop like it would have before  if it was outside of the conditional statement and the configuration blocks were combined.